+++ 1.16_combat-6/src/main/java/io/github/axolotlclient/modules/hud/HudManager.java
@@ -118,6 +118,7 @@ public class HudManager extends AbstractModule {
 		add(new ComboHud());
 		add(new PlayerHud());
 		add(new MouseMovementHud());
+		add(new DebugCountersHud());
 		entries.put(BedwarsMod.getInstance().getUpgradesOverlay().getId(), BedwarsMod.getInstance().getUpgradesOverlay());
 
 		entries.values().forEach(HudEntry::init);