+++ 1.8.9/src/main/java/io/github/axolotlclient/modules/hud/HudManager.java
@@ -109,6 +109,7 @@ public class HudManager extends AbstractModule {
 		add(new PlayerHud());
 		add(new ChatHud());
 		add(new MouseMovementHud());
+		add(new DebugCountersHud());
 		entries.put(BedwarsMod.getInstance().getUpgradesOverlay().getId(), BedwarsMod.getInstance().getUpgradesOverlay());
 
 		entries.values().forEach(HudEntry::init);